requirements.txt
openai==0.27.2
flet==0.4.2
python-dotenv==1.0.0